code/requirements.txt
numpy
hydra-core
torch
diffusers
transformers
accelerate
safetensors
Pillow
torchvision
torchcodec
ftfy
scipy
tqdm
